add getRandomInt & getRandomListNumbers test, move localStorage funtions
