+++ helpers/utils.tsx
@@ -118,24 +118,3 @@ export const getTitleText = (total: number) => ({
   totalText: total > 0 ? total : 'There are no',
   restaurantText: `restaurant${total === 0 || total >= 2 ? 's' : ''}`,
 });
-
-export const loadStateFromLocalStorage = () => {
-  try {
-    const serializedState = localStorage.getItem('state');
-    if (serializedState === null) {
-      return undefined;
-    }
-    return JSON.parse(serializedState);
-  } catch (error) {
-    return undefined;
-  }
-};
-
-export const saveStateToLocalStorage = (state: any) => {
-  try {
-    const serializedState = JSON.stringify(state);
-    localStorage.setItem('state', serializedState);
-  } catch (error) {
-    console.log(`${error} on trying to save global state in LocalStorage`);
-  }
-};
